refactor: move clock
